added import from xlsx files

+++ controllers/plugin_lookout.py
@@ -16,6 +16,8 @@
 #    You should have received a copy of the GNU General Public License
 #    along with Plugin_lookout.  If not, see <http://www.gnu.org/licenses/>.
 
+#from random import randint
+
 def index(): return dict()
 
 #        # condizione: esiste tabella con lo stesso nome in una connessione con stesso alias
@@ -66,11 +68,10 @@ def connection_onvalidation(form):
 def connection_oncreate(form):
     set_data_permission('plugin_lookout_connections', form.vars.id)
 
-
 @auth.requires_login()
 def plugin_lookout_connections():
-    db.plugin_lookout_connections.dsn.represent = lambda v, r: '%s: %s' % (r.alias, v.replace('%s', '<password>'))
     
+#    db.plugin_lookout_connections.dsn.represent = lambda v, r: '%s: %s' % (r.alias, v.replace('%s', randint(8, 16)*'*'))
     form = SQLFORM.smartgrid(db.plugin_lookout_connections,
         onvalidation = connection_onvalidation,
         linked_tables = [],
@@ -104,27 +105,31 @@ def table_onvalidate(form):
     # before deletion
     if form.vars.delete_this_record=='on':
 #        join = db.plugin_lookout_tables.connection_id==db.plugin_lookout_connections.id
-        where = db.plugin_lookout_tables.id==form.vars.id
+        where = db.plugin_lookout_tables.id==request.vars.id
         rec_table = db(where).select(
             db.plugin_lookout_tables.table_migrate,
             db.plugin_lookout_tables.is_view,
+            db.plugin_lookout_tables.is_active,
             db.plugin_lookout_tables.connection_name
         ).first()
-        table_migrate = rec_table.migrate
+        table_migrate = rec_table.table_migrate
         table_is_view = rec_table.is_view
         connection_name = rec_table.connection_name
+        table_is_active = rec_table.is_active
         
         key = '%s_%s' % (globals()[connection_name]._uri_hash, form.vars.table_name)
         db(db.auth_group.role.contains(key)).delete()
         
-        for row in db(db.plugin_lookout_fields.tables.contains(form.vars.id)).select():
-            ids = row.tables
-            ids.remove(int(form.vars.id))
-            row.update_record(tables=ids)
+        db(db.plugin_lookout_fields.table_id==request.vars.id).delete()
+        
+#        for row in db(db.plugin_lookout_fields.tables.contains(form.vars.id)).select():
+#            ids = row.tables
+#            ids.remove(int(form.vars.id))
+#            row.update_record(tables=ids)
             
         if not table_migrate and table_is_view:
             db.executesql('DROP VIEW %s CASCADE;' % tab.table_name)
-        elif table_migrate:
+        elif table_migrate and table_is_active:
             try:
                 globals()[connection.alias][form.vars.table_name].drop()
             except Exception, error:
@@ -134,16 +139,17 @@ def table_onvalidate(form):
                 form.errors.external_error = str(msg)
         
 def table_oncreate(form):
-    if 'connection_id' in form.vars:
-        connection_name = db.plugin_lookout_connections[form.vars.connection_id].alias
-    else:
-        connection_name = db.plugin_lookout_tables[form.vars.id].connection_name
-    if form.vars.restricted:
-        t_hash = '%s_%s' % (globals()[connection_name]._uri_hash, form.vars.table_name)
-        w_role = 'write_%s' % t_hash
-        r_role = 'read_%s' % t_hash
-        set_data_permission('plugin_lookout_tables', form.vars.id, role=r_role)
-        set_data_permission('plugin_lookout_tables', form.vars.id, role=w_role)
+    if form.vars.is_active:
+        if 'connection_id' in form.vars:
+            connection_name = db.plugin_lookout_connections[form.vars.connection_id].alias
+        else:
+            connection_name = db.plugin_lookout_tables[form.vars.id].connection_name
+        if form.vars.restricted:
+            t_hash = '%s_%s' % (globals()[connection_name]._uri_hash, form.vars.table_name)
+            w_role = 'write_%s' % t_hash
+            r_role = 'read_%s' % t_hash
+            set_data_permission('plugin_lookout_tables', form.vars.id, role=r_role)
+            set_data_permission('plugin_lookout_tables', form.vars.id, role=w_role)
         
 def table_ondelete():
     unusefull_fields_set = db(db.plugin_lookout_fields.tables==[])
@@ -157,7 +163,6 @@ def plugin_lookout_tables():
     * aggiungi campo a tabella
     * rimuovi campo dalla tabella
     '''
-    from random import randint
     
     if 'edit' in request.args:
         db.plugin_lookout_tables.table_name.writable = False
@@ -166,14 +171,17 @@ def plugin_lookout_tables():
     only_view = bool(request.vars.only_view)
     session.plugin_lookout_tables_only_view = only_view
     if 'edit' in request.vars:
-        editable = auth.has_permission('plugin_lookout_tables', request.args(0), auth.user_id)
+        editable = auth.has_permission('plugin_lookout_tables',
+            request.args(0),
+            auth.user_id
+        )
     else:
         editable = not session.plugin_lookout_tables_only_view
 
     table_set = get_table_set(view_only=bool(only_view))
 
     db.plugin_lookout_tables.table_name.represent = lambda val,row: A(row.table_name, _href=URL('plugin_lookout_external_tables', vars=dict(table_id=row.id)))
-    db.plugin_lookout_tables.connection_id.represent = lambda val,row: ('%(dsn)s' % db.plugin_lookout_connections[val]).replace('%s', randint(8, 16)*'*')
+#    db.plugin_lookout_tables.connection_id.represent = lambda val,row: ('%(dsn)s' % db.plugin_lookout_connections[val]).replace('%s', randint(8, 16)*'*')
     form = SQLFORM.smartgrid(db.plugin_lookout_tables,
         onvalidation = table_onvalidate,
         oncreate = table_oncreate,
@@ -185,14 +193,19 @@ def plugin_lookout_tables():
         constraints = dict(plugin_lookout_tables=table_set) # get_data_with_permissions('plugin_lookout_tables')
     )
     
+    plugin_lookout_tables_menu = None
     if 'view' in request.args:
         table_id = request.args[-1]
-        if db((db.auth_permission.table_name=='plugin_lookout_tables')\
-            &(db.auth_permission.record_id==table_id)).count():
-            my_extra_element = FORM(INPUT(_type="submit", _value=T('Share data')), _action=URL('share_data_with_users', args=table_id))
-            form.components.insert(-1,my_extra_element)
-    
-    return dict(form=form)
+        plugin_lookout_tables_menu = [
+            (T('Share data'), False, URL('share_data_with_users', args=table_id), []),
+            (T('Edit table fields'), False, URL('plugin_lookout_fields',
+                vars=dict(keywords='plugin_lookout_fields.table_id="%s"' % table_id)), [])
+        ]
+        if db(db.plugin_lookout_datafiles.table_id==table_id).count():
+            plugin_lookout_tables_menu += [
+                (T('Init/Reset table data'), False, URL('init_xls_data', vars=dict(table_id=table_id)), []),
+            ]
+    return dict(form=form, plugin_lookout_tables_menu=plugin_lookout_tables_menu)
 
 ######################################################## PLUGIN_LOOKOUT_FIELDS #
 
@@ -203,8 +216,10 @@ def plugin_lookout_tables():
 #    ) for i in db(db.plugin_lookout_tables.id.belongs(id))\
 #        .select(db.plugin_lookout_tables.id, db.plugin_lookout_tables.table_name)])
 
-@auth.requires_login()
+@auth.requires(control_permission(request.vars.table_id, reading=False, default=True), requires_login=True)
 def plugin_lookout_fields():
+
+    table_condition = get_table_set(view_only=False)
     db.plugin_lookout_fields.table_id.requires = IS_IN_DB(db(get_table_set(view_only=False)), db.plugin_lookout_tables.id, '%(table_name)s')
     db.plugin_lookout_fields.table_id.represent = lambda value, row: db.plugin_lookout_tables[value].table_name
     form = SQLFORM.smartgrid(db.plugin_lookout_fields,
@@ -213,9 +228,9 @@ def plugin_lookout_fields():
 #        editable=False,
 #        deletable=False,
         constraints = dict(plugin_lookout_fields=(db.plugin_lookout_fields\
-            .table_id.belongs(db(get_table_set(view_only=False))._select(db.plugin_lookout_tables.id))))
+            .table_id.belongs(db(table_condition)._select(db.plugin_lookout_tables.id))))
     )
-    return locals()
+    return dict(form=form)
 
 ####################################################################### TABLES #
 
@@ -225,9 +240,9 @@ def plugin_lookout_external_tables():
     Controller for manage data inside table that are not part of the model
     '''
     message = 'Here you can see the date inside the tables you have configured'
+
     table_id = request.vars.table_id or session.plugin_lookout_external_tables_id or redirect(URL('plugin_lookout_tables'))
     session.plugin_lookout_external_tables_id = table_id
-#    table_id = int(request.args(0)) or redirect(URL('plugin_lookout_tables'))
     
     check_message = IS_IN_DB(db, 'plugin_lookout_tables.id')(table_id)[1]
     if check_message:
@@ -296,6 +311,150 @@ def share_data_with_users():
 #        table = globals()[rec_table.plugin_lookout_connections.alias][rec_table.plugin_lookout_tables.table_name]
 #        share_data(table, read_only=form.vars.read_only, users=users)
     return dict(form=form)
+    
+
+################################################################### XLS IMPORT #
+
+from openpyxl.reader.excel import load_workbook
+from plugin_lookout import guess_type
+
+@auth.requires_login()
+def import_xls_data():
+
+    table_id = request.vars.table_id or redirect(URL('plugin_lookout_tables'))
+    form = SQLFORM.factory(
+        Field('source_file', 'upload', uploadfolder=uploadfolder)
+    )
+    if form.accepts(request, session):
+        filePath = os.path.join(uploadfolder, form.vars.source_file)
+        xlsx = load_workbook(filePath)
+        first_sheet_name = xlsx.get_sheet_names()[0]
+        sheet = xlsx.get_sheet_by_name(first_sheet_name)
+        pass
+
+    return dict()
+
+#@auth.requires_login()
+#def import_csv_upload_file():
+#    '''
+#    dict = {"<head>": dict(
+#        field_name = <>,
+#        field_type = <>,
+#    )}
+#    '''
+#    
+#    form = SQLFORM.factory(
+#        db.plugin_lookout_tables,
+#        Field('csv_file', 'upload', uploadfolder=uploadfolder)
+#    )
+#    if form.accepts(request, session):
+#        path = os.path.join(uploadfolder, form.vars.csv_file)
+#        new_tab_record = dict([(i, form.vars[i]) for i in form.vars if i in db.plugin_lookout_tables.fields])        
+
+#    return dict(form=form)
+
+#def csv_import_customize_field():
+#    
+#    return dict()
+#    
+    
+@auth.requires_login()
+def import_xls():
+    '''
+    This controller is for build a table structure for containing data from an
+    excell spread sheet. It creates ad many fields as the number of the columns
+    if finds in the first sheet of the xls file and try to guess the type of
+    data to contain for each field. The table structure is created not active so
+    you can made some change before to import data.
+    '''
+
+    db.plugin_lookout_tables.is_active.default = False
+    db.plugin_lookout_tables.is_active.writable = False
+    if not 'new' in request.args: redirect(URL('plugin_lookout_tables'))
+
+    form = SQLFORM.factory(
+        db.plugin_lookout_tables,
+        db.plugin_lookout_datafiles
+#        Field('source_file', 'upload', uploadfolder=uploadfolder)
+    )
+
+    if form.accepts(request, session, onvalidation=table_onvalidate):
+    
+        table_id = db.plugin_lookout_tables.insert(**db.plugin_lookout_tables._filter_fields(form.vars))
+        form.vars.table_id=table_id
+        file_id = db.plugin_lookout_datafiles.insert(**db.plugin_lookout_datafiles._filter_fields(form.vars))
+    
+        filePath = os.path.join(uploadfolder, db.plugin_lookout_datafiles[table_id].file_name)
+        
+        xlsx = load_workbook(filePath)
+        first_sheet_name = xlsx.get_sheet_names()[0]
+        sheet = xlsx.get_sheet_by_name(first_sheet_name)
+        header = sheet.rows[0]
+
+        for idx, cell in enumerate(header):
+            values = [i.value for i in sheet.columns[idx]][1:]
+            
+            field_type = guess_type(values)
+            
+            ret = db.plugin_lookout_fields.validate_and_insert(
+                table_id = table_id,
+                field_name = cell.value.lower().replace(' ', ''),
+                field_comment = cell.value,
+                field_type = field_type
+            )
+            if ret.get('error'):
+                db.plugin_lookout_fields.insert(
+                    table_id = table_id,
+                    field_name = 'field_%s' % idx,
+                    field_label = cell.value,
+                    field_comment = str(ret.error),
+                    field_type = field_type
+                )
+        xlsx.close()
+        redirect(URL('plugin_lookout_fields',
+                vars=dict(keywords='plugin_lookout_fields.table_id="%s"' % table_id)))
+    
+    return dict(form=form)
+
+@auth.requires_login()
+def init_xls_data():
+    table_id = request.vars.table_id or redirect(URL('plugin_lookout_tables'))
+    
+    table_info = db.plugin_lookout_tables[table_id]
+    
+    if not table_info.is_active:
+        session.flash = T('You cannot populate not active tables')
+        redirect(URL('plugin_lookout_tables', args=['plugin_lookout_tables', 'edit', 'plugin_lookout_tables', table_id]))
+    
+    filePath = os.path.join(uploadfolder, db.plugin_lookout_datafiles[table_id].file_name)
+    xlsx = load_workbook(filePath)
+    first_sheet_name = xlsx.get_sheet_names()[0]
+    sheet = xlsx.get_sheet_by_name(first_sheet_name)
+    
+    res_fields = db(db.plugin_lookout_fields.table_id==table_id).select(db.plugin_lookout_fields.field_name)
+    
+    field_names = [i.field_name for i in res_fields]
+    
+    mydb = globals()[table_info.connection_name]
+    error = None
+    for row in sheet.rows[1:]:
+        values = [cell.value for cell in row]
+        
+        kwargs = dict([(k,v) for k,v in zip(field_names, values)])
+        ret = mydb[table_info.table_name].validate_and_insert(**kwargs)
+        if ret.errors:
+            error = T('Import error in line %(line)s, column %(key)s: %(value)s (%(data)s)',dict(line=n, data=data, **ret.errors))
+            mydb.rollback()
+            break
+    
+    if error:
+        session.flash = error
+        redirect(URL('plugin_lookout_tables'))
+    
+    redirect(URL('plugin_lookout_external_tables', vars=dict(table_id=table_id)))
+        
+    
+
